requirements.txt
openai>=1.60.0
openpyxl>=3.1.0
xlrd>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
watchdog>=4.0.0
thefuzz>=0.22.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
gemini-webapi>=1.21.0
